# perf(mempool): use atomic instead of mutex to improve perf