---
Use mutex
```
BenchmarkCurrentRedundancy
BenchmarkCurrentRedundancy/CurrentRedundancy
BenchmarkCurrentRedundancy/CurrentRedundancy-11             172358114            6.857 ns/op
BenchmarkCurrentRedundancy/IncFirstTimeTxs
BenchmarkCurrentRedundancy/IncFirstTimeTxs-11               227235248            5.264 ns/op
BenchmarkCurrentRedundancy/IncDuplicateTxs
BenchmarkCurrentRedundancy/IncDuplicateTxs-11               227110046            5.264 ns/op

```

Use atomic
```
BenchmarkCurrentRedundancy
BenchmarkCurrentRedundancy/CurrentRedundancy
BenchmarkCurrentRedundancy/CurrentRedundancy-11             1000000000           0.4400 ns/op
BenchmarkCurrentRedundancy/IncFirstTimeTxs
BenchmarkCurrentRedundancy/IncFirstTimeTxs-11               592631368            1.848 ns/op
BenchmarkCurrentRedundancy/IncDuplicateTxs
BenchmarkCurrentRedundancy/IncDuplicateTxs-11               665309712            1.772 ns/op
```

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use [unclog](https://github.com/informalsystems/unclog) to manage our changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code comments